Rename templateString in RuntimeNodeDetailsV2 to pipelineJobString Extract 'kubernetes' as a constant. Put getComponentSpec in a new file called NodeUtils
