+++ frontend/src/pages/RunDetailsV2.tsx
@@ -213,7 +213,7 @@ export function RunDetailsV2(props: RunDetailsV2Props) {
                 <RuntimeNodeDetailsV2
                   layers={layers}
                   onLayerChange={layerChange}
-                  templateString={pipelineJobStr}
+                  pipelineJobString={pipelineJobStr}
                   runId={runId}
                   element={selectedNode}
                   elementMlmdInfo={selectedNodeMlmdInfo}
